moved wsi base class to backends folder
