+++ src/eva/vision/data/wsi/backends/base.py
